+++ eio-okta-sync/src/command/authentik.rs
@@ -1,5 +1,3 @@
-use crate::authentik::traits::RunAsync;
-
 #[derive(Debug, Clone, clap::Args)]
 #[command(about = "experimental support for Authentik")]
 pub struct Command {
@@ -9,6 +7,6 @@ pub struct Command {
 
 impl Command {
   pub(crate) async fn run(self) -> Result<(), crate::Error> {
-    Ok(self.subcommand.run().await?)
+    self.subcommand.run().await
   }
 }